+++ app/src/main/kotlin/ac/mdiq/podcini/storage/database/RealmDB.kt
@@ -40,11 +40,9 @@ object RealmDB {
                 SubscriptionLog::class,
                 Chapter::class,
                 PAFeed::class,
-                AStream::class,
-                VStream::class,
             ))
             .name("Podcini.realm")
-            .schemaVersion(41)
+            .schemaVersion(42)
             .migration({ mContext ->
                 val oldRealm = mContext.oldRealm // old realm using the previous schema
                 val newRealm = mContext.newRealm // new realm using the new schema